image_downloader/requirements.txt
pillow
requests
